split out the codec
