+++ tests/subscriptions/test_store.py
@@ -1,13 +1,8 @@
-import json
 from datetime import timedelta
 
 from snuba.redis import redis_client
 from snuba.subscriptions.data import Subscription
-from snuba.subscriptions.store import (
-    RedisSubscriptionStore,
-    SubscriptionCodec,
-)
-from tests.base import BaseTest
+from snuba.subscriptions.store import RedisSubscriptionStore
 from tests.subscriptions import BaseSubscriptionTest
 
 
@@ -78,44 +73,3 @@ class TestRedisSubscriptionStore(BaseSubscriptionTest):
         store_2.create(new_subscription_id, new_subscription)
         assert store_1.all() == [(subscription_id, self.subscription)]
         assert store_2.all() == [(new_subscription_id, new_subscription)]
-
-
-class TestSubscriptionCodec(BaseTest):
-    def build_subscription(self) -> Subscription:
-        return Subscription(
-            project_id=5,
-            conditions=[["platform", "IN", ["a"]]],
-            aggregations=[["count()", "", "count"]],
-            time_window=timedelta(minutes=500),
-            resolution=timedelta(minutes=1),
-        )
-
-    def test_basic(self):
-        subscription = self.build_subscription()
-        codec = SubscriptionCodec()
-        assert codec.decode(codec.encode(subscription)) == subscription
-
-    def test_encode(self):
-        codec = SubscriptionCodec()
-        subscription = self.build_subscription()
-
-        payload = codec.encode(subscription)
-        data = json.loads(payload.decode("utf-8"))
-        assert data["project_id"] == subscription.project_id
-        assert data["conditions"] == subscription.conditions
-        assert data["aggregations"] == subscription.aggregations
-        assert data["time_window"] == int(subscription.time_window.total_seconds())
-        assert data["resolution"] == int(subscription.resolution.total_seconds())
-
-    def test_decode(self):
-        codec = SubscriptionCodec()
-        subscription = self.build_subscription()
-        data = {
-            "project_id": subscription.project_id,
-            "conditions": subscription.conditions,
-            "aggregations": subscription.aggregations,
-            "time_window": int(subscription.time_window.total_seconds()),
-            "resolution": int(subscription.resolution.total_seconds()),
-        }
-        payload = json.dumps(data).encode("utf-8")
-        assert codec.decode(payload) == subscription
